* fix bug delete.sh 修复删除N天前文件 如果找不到文件 * 会列出当前路径下所有文件一bug

+++ shell/util/delete.sh
@@ -8,7 +8,7 @@
 ## 支持 -o 参数 将删除的文件放取指定日志文件中方便其它脚本调用写日志
 
 author=man003@163.com
-version=V1.3-20161208
+version=V1.3-20171107
 
 #==============================TODO==============================
 # -n参数不支持文件名带有空格
@@ -116,6 +116,8 @@ exit 0
 	# * 支持version/outputRntime参数 把原来的-V选项去掉
 # 20161208 V1.3
 	# * -r选项可单独使用删除空目录 原来必须和-d -n选项同时使用才行 
+# 20171107 V1.3-20171107
+   # * fix bug 解决删除N天前文件 如果未找到文件 则会列出当前路径下所有文件的bug 
 
 #########################如下是配置区域#########################################################
 
@@ -320,15 +322,23 @@ then
 	#echo "find $localPath -name "*$suffix" -mtime +"$deleteDays"|xargs ls -l";
 	# 注不要写成 find . -name *.* 这样只能查temp.out这样的文件 但查不出temp这样的文件
 	# 如下find -print0|xargs -0 是为了处理文件名有空格的文件
-    # TODO-BUG  如果find未查出文件 则会传一个空给ls -l 则其会列出所有的文件
-    # 虽然实际并没有删除文件 因为 rm -rf 不加参数什么也删除不了,但写的日志不对
-	find $deletePath -name "*$suffix" -mtime +"$deleteDays" -print0|xargs -0 ls -l|tee $deleteTmp;
 
-	# 如果输入-l参数 则只列出要删除的内容 不实际删除 免得删除错了
-	if [ ! ${IS_DEBUG}X = "true"X  ]
-	then
-		find $deletePath -name "*$suffix" -mtime +"$deleteDays" -print0|xargs -0 -n5 rm -rf;
-	fi
+    # 先判断下是否找到文件 如果找到了再输出 主要是解决如下的bug
+    # 如果find未查出文件 则会传一个空给ls -l 则其会列出当前路径下所有的文件
+    # 虽然实际并没有删除文件 因为 rm -rf 不加参数什么也删除不了,但写的日志不对会对你实际运行时产生误导
+    findResult=`find $deletePath -name "*$suffix" -mtime +"$deleteDays" -print0`
+    if [ ! -z $findResult ]
+    then
+        find $deletePath -name "*$suffix" -mtime +"$deleteDays" -print0|xargs -0 ls -l|tee $deleteTmp;
+        # 如果输入-l参数 则只列出要删除的内容 不实际删除 免得删除错了
+        if [ ! ${IS_DEBUG}X = "true"X  ]
+        then
+            find $deletePath -name "*$suffix" -mtime +"$deleteDays" -print0|xargs -0 -n5 rm -rf;
+        fi
+    else
+        echo "无可删除的文件"
+    fi
+
 elif [ ${shellFunction}X = "retainNewest"X ]
 then
 	#echo "retainNewest" 功能";
